nsc-events-nextjs_9_537_signin-change-forgot-pw-responsive-ui
Resolves #537

This PR ensures that the Sign In, Change and Forgot Password pages has responsive UI and adjusts to mobile screen sizes well, with no cutoff or scrolling horizontally. I also made improvements to the appearance of the pages for tablet and desktop screen sizes.

Test this PR by right clicking the page and using Inspect elements to enter the Toggle Device Toolbar and adjust the screen sizes.